packages/benchmark-py/pyproject.toml
[project]
name = "engram-benchmark"
version = "0.1.0"
description = "LongMemEval benchmark suite for Engram memory system"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
	{ name = "Engram Team" }
]

dependencies = [
	# CLI & Terminal UI
	"typer>=0.15.0",
	"rich>=13.9.0",
	"click>=8.1.0",

	# Config & Validation
	"pydantic>=2.10.0",
	"pydantic-settings>=2.6.0",
	"pyyaml>=6.0.0",

	# LLM
	"litellm>=1.55.0",
	"anthropic>=0.42.0",
	"openai>=1.57.0",

	# Embeddings
	"sentence-transformers>=3.3.0",
	"torch>=2.5.0",

	# Vector DB
	"qdrant-client>=1.12.0",
	"chromadb>=0.5.0",

	# Metrics
	"ranx>=0.3.21",
	"ragas>=0.2.0",
	"scikit-learn>=1.6.0",
	"numpy>=1.26.0",

	# Data
	"datasets>=3.2.0",
	"pandas>=2.2.0",

	# Temporal
	"python-dateutil>=2.9.0",
	"dateparser>=1.2.0",

	# HTTP
	"httpx>=0.28.0",

	# Progress
	"tqdm>=4.67.0",
]

[project.optional-dependencies]
mteb = [
	"mteb>=1.35.0",
	"beir>=2.2.0",
]

dev = [
	"pytest>=8.3.0",
	"pytest-asyncio>=0.24.0",
	"pytest-cov>=6.0.0",
	"ruff>=0.8.0",
	"mypy>=1.13.0",
	"types-python-dateutil>=2.9.0",
	"types-pyyaml>=6.0.0",
]

[project.scripts]
engram-benchmark = "engram_benchmark.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/engram_benchmark"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
	"E",   # pycodestyle errors
	"W",   # pycodestyle warnings
	"F",   # pyflakes
	"I",   # isort
	"N",   # pep8-naming
	"UP",  # pyupgrade
	"B",   # flake8-bugbear
	"C4",  # flake8-comprehensions
	"SIM", # flake8-simplify
	"TCH", # flake8-type-checking
]
ignore = [
	"E501",  # line too long (handled by formatter)
	"B008",  # do not perform function calls in argument defaults
	"B006",  # do not use mutable data structures for argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["engram_benchmark"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
	"litellm.*",
	"ranx.*",
	"ragas.*",
	"datasets.*",
	"sentence_transformers.*",
	"chromadb.*",
	"qdrant_client.*",
	"dateparser.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
	"--strict-markers",
	"--strict-config",
	"-ra",
]
